pyproject.toml
[project]
name = "holiday-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.6",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.28.0",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.10",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "uvicorn-worker>=0.3.0",
]

[tool.setuptools.packages.find]
where = ["src"]
